+++ CHANGES.md
@@ -3,6 +3,9 @@ NoFlo ChangeLog
 
 ## 1.3.0 (git master)
 
+* NoFlo `createNetwork` now defaults to the non-legacy "network drives graph" mode
+* NoFlo `createNetwork` now only supports the `graph, options, callback` signature, no options given in some other order
+* `noflo.Network` interface has been removed. Use `createNetwork` to instantiate networks
 * CoffeeScript is no longer bundled with NoFlo. Install the CoffeeScript compiler in your project if you need to be able to load CoffeeScript components
 
 ## 1.2.7 (November 13th 2020)